Allow custom auth backend mount point
